Refactor processor encoding methods into separate module
New encoding module is nearly identical to the client's encoding module,
and is ready to be split into problems/solutions.
